+++ src/1 - AvanadeEstudosAPI.API/ViewModels/UpdateUserViewModel.cs
@@ -5,8 +5,8 @@ namespace AvanadeEstudosAPI.API.ViewModels
     public class UpdateUserViewModel
     {
         [Required(ErrorMessage = "O Id não pode ser vazio.")]
-        [Range(1, long.MaxValue, ErrorMessage = "O id não pode ser menor que 1.")]
-        public long Id { get; set; }
+        [Range(1, int.MaxValue, ErrorMessage = "O id não pode ser menor que 1.")]
+        public int Id { get; set; }
 
         [Required(ErrorMessage = "O nome não pode ser vazio.")]
         [MinLength(3, ErrorMessage = "O nome deve ter no mínimo 3 caracteres.")]
