+++ src/2 - AvanadeEstudosAPI.Domain/Interfaces/IBaseRepository.cs
@@ -1,15 +1,12 @@
-using System.Threading.Tasks;
 using AvanadeEstudosAPI.Domain.Entities;
-using System.Collections.Generic;
-using System.Linq.Expressions;
-using System;
 
-namespace AvanadeEstudosAPI.Domain.Interfaces{
+namespace AvanadeEstudosAPI.Domain.Interfaces
+{
     public interface IBaseRepository<T> where T : Base{
         Task<T> CreateAsync(T obj);
         Task<T> UpdateAsync(T obj);
-        Task RemoveAsync(long id);
+        Task RemoveAsync(int id);
         Task<List<T>> GetAllAsync();
-        Task<T> GetAsync(long id);        
+        Task<T> GetAsync(int id);        
     }
 }
