+++ src/3 - AvanadeEstudosAPI.Services/DTO/UserDTO.cs
@@ -3,15 +3,14 @@ namespace AvanadeEstudosAPI.Services.DTO
     public class UserDTO
     {
          //Propriedades
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string Name { get; set; }
         public string Email { get; set; }
-        public string Password { get; set; }
-        
+        public string Password { get; set; }        
 
         public UserDTO(){}
        
-        public UserDTO(long id, string name, string email, string password)
+        public UserDTO(int id, string name, string email, string password)
         {
             Id = id;
             Name = name;
